Glueing mulltiple plates from a folder

This function takes a folder path and on the available .xlsx files the revive_xfplate() function is run and output in one nested tibble.

Usage

glue_xfplates(folderpath_seahorse, arg_is_folder)

Arguments

folderpath_seahorse

the path to a folder where the .xlsx files are located or a vector of strings pointing to the path of each individual file

arg_is_folder

either TRUE or FALSE. When the input is a vector of path strings use FALSE, is it points to a folder use TRUE

Value

a nested tibble with all files organized in a row

See also

Examples

c(system.file("extdata",
  "20191219_SciRep_PBMCs_donor_A.xlsx", 
  package = "seahtrue"),
  system.file("extdata",
  "20200110_SciRep_PBMCs_donor_B.xlsx", 
  package = "seahtrue"),
  system.file("extdata",
  "20200110_SciRep_PBMCs_donor_C.xlsx", 
  package = "seahtrue")) |> 
  glue_xfplates(arg_is_folder = FALSE)
#> → Start function to read seahorse plate data from Excel file:
#> 20191219_SciRep_PBMCs_donor_A.xlsx
#> ℹ Finished collecting assay information.
#> → plateid is identified as:V0174416419V
#> → Rate was exported WITH background correction
#> ℹ Finished preprocessing of the input data
#> ℹ Finished validating the input data
#> → Start function to read seahorse plate data from Excel file:
#> 20200110_SciRep_PBMCs_donor_B.xlsx
#> ℹ Finished collecting assay information.
#> → plateid is identified as:V0194115019V
#> → Rate was exported WITH background correction
#> ℹ Finished preprocessing of the input data
#> ℹ Finished validating the input data
#> → Start function to read seahorse plate data from Excel file:
#> 20200110_SciRep_PBMCs_donor_C.xlsx
#> ℹ Finished collecting assay information.
#> → plateid is identified as:V0194715019V
#> → Rate was exported WITH background correction
#> ℹ Finished preprocessing of the input data
#> ℹ Finished validating the input data
#> # A tibble: 3 × 9
#>   plate_id     filepath_seahorse date_run         date_processed      assay_info
#>   <chr>        <list>            <chr>            <dttm>              <list>    
#> 1 V0174416419V <tibble [1 × 3]>  19-12-2019 17:25 2024-04-19 13:34:02 <tibble>  
#> 2 V0194115019V <tibble [1 × 3]>  10-1-2020 16:59  2024-04-19 13:34:07 <tibble>  
#> 3 V0194715019V <tibble [1 × 3]>  15-1-2020 17:26  2024-04-19 13:34:11 <tibble>  
#> # ℹ 4 more variables: injection_info <list>, raw_data <list>, rate_data <list>,
#> #   validation_output <list>