Running the read, validate and preprocess

This function takes the Seahorse Wave excel file and computes it through read, validate and preprocess

Usage

run_seahtrue(filepath_seahorse, ...)

Arguments

filepath_seahorse

Absolute path to the Seahorse Excel file.

...

Dots argument (…) allows the function to take an undefined number of arguments.

Value

A preprocessed seahorse dataset is returned as an output. This is a nested tibble with the following 7 columns:

* plate_id = Barcode from the 'Assay Configuration' sheet from the Seahorse Excel file.

* filepath_seahorse = Path to Seahorse Excel file.

* date = Date derived from 'Assay Configuration' sheet of the Seahorse Excel file.

* assay_info = Dataframe with information from the 'Assay Configuration' sheet from the Seahorse Excel file.

* injection_info = Dataframe with information derived from 'Operation log' sheet.

* raw_data = Preprocessed raw dataframe from Seahorse 'Raw' sheet.

* rate_data = Preprocessed rate data from 'Rate' sheet.

Examples

run_seahtrue(system.file("extdata", "20191219_SciRep_PBMCs_donor_A.xlsx", package = "seahtrue"))
#> Starting Seahtrue analysis
#> ℹ Preprocessing collected seahorse information.
#> → Start function to read seahorse plate data from Excel file:
#> /private/var/folders/27/brlq29b170xfkshv_mmttv3c0000gn/T/Rtmp4vN3qr/temp_libpathe9e54d64d368/seahtrue/extdata/20191219_SciRep_PBMCs_donor_A.xlsx
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> ℹ Finished collecting assay information.
#> ℹ Finished preprocessing collected seahorse information.
#> ℹ Validate the preprocessed data.
#> # A tibble: 1 × 8
#>   plate_id  filepath_seahorse date  assay_info injection_info raw_data rate_data
#>   <chr>     <list>            <chr> <list>     <list>         <list>   <list>   
#> 1 V0174416… <tibble [1 × 3]>  19-1… <tibble>   <tibble>       <tibble> <tibble> 
#> # ℹ 1 more variable: validation_failures <list>
run_seahtrue(system.file("extdata", "20200110_SciRep_PBMCs_donor_B.xlsx", package = "seahtrue"))
#> Starting Seahtrue analysis
#> ℹ Preprocessing collected seahorse information.
#> → Start function to read seahorse plate data from Excel file:
#> /private/var/folders/27/brlq29b170xfkshv_mmttv3c0000gn/T/Rtmp4vN3qr/temp_libpathe9e54d64d368/seahtrue/extdata/20200110_SciRep_PBMCs_donor_B.xlsx
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> ℹ Finished collecting assay information.
#> ℹ Finished preprocessing collected seahorse information.
#> ℹ Validate the preprocessed data.
#> # A tibble: 1 × 8
#>   plate_id  filepath_seahorse date  assay_info injection_info raw_data rate_data
#>   <chr>     <list>            <chr> <list>     <list>         <list>   <list>   
#> 1 V0194115… <tibble [1 × 3]>  10-1… <tibble>   <tibble>       <tibble> <tibble> 
#> # ℹ 1 more variable: validation_failures <list>
run_seahtrue(system.file("extdata", "20200110_SciRep_PBMCs_donor_C.xlsx", package = "seahtrue"))
#> Starting Seahtrue analysis
#> ℹ Preprocessing collected seahorse information.
#> → Start function to read seahorse plate data from Excel file:
#> /private/var/folders/27/brlq29b170xfkshv_mmttv3c0000gn/T/Rtmp4vN3qr/temp_libpathe9e54d64d368/seahtrue/extdata/20200110_SciRep_PBMCs_donor_C.xlsx
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> New names:
#> • `` -> `...1`
#> ℹ Finished collecting assay information.
#> ℹ Finished preprocessing collected seahorse information.
#> ℹ Validate the preprocessed data.
#> # A tibble: 1 × 8
#>   plate_id  filepath_seahorse date  assay_info injection_info raw_data rate_data
#>   <chr>     <list>            <chr> <list>     <list>         <list>   <list>   
#> 1 V0194715… <tibble [1 × 3]>  15-1… <tibble>   <tibble>       <tibble> <tibble> 
#> # ℹ 1 more variable: validation_failures <list>
if (FALSE) {
run_seahtrue(here::here("inst", "extdata", "20191219_SciRep_PBMCs_donor_A.xlsx"))
run_seahtrue(here::here("inst", "extdata", "20200110_SciRep_PBMCs_donor_B.xlsx"))
run_seahtrue(here::here("inst", "extdata", "20200110_SciRep_PBMCs_donor_C.xlsx"))
}